(tae): txn refactor 6
