+++ pkg/vm/engine/tae/tables/block.go
@@ -13,6 +13,7 @@ import (
 	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/container/vector"
 	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/index/access/acif"
 	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/index/access/impl"
+	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/tables/jobs"
 	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/tables/updates"
 	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/tasks"
 	"github.com/matrixorigin/matrixone/pkg/vm/engine/tae/txn/txnbase"
@@ -137,7 +138,7 @@ func (blk *dataBlock) EstimateScore() int {
 }
 
 func (blk *dataBlock) BuildCheckpointTaskFactory() (factory tasks.TxnTaskFactory, err error) {
-	factory = CompactBlockTaskFactory(blk.meta)
+	factory = jobs.CompactBlockTaskFactory(blk.meta)
 	return
 	// if !blk.meta.IsAppendable() {
 	// }
